move custom analyses out and disable them
